+++ src/controllers/event.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import response from "../utils/response";
 import { eventDAO, EventModel, TEvent } from "../models/event.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 export default {
   async create(req: IReqUser, res: Response) {
@@ -61,34 +61,48 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notfound(res, "failed to find one an event");
+      }
       const result = await EventModel.findById(id);
-      response.success(res, result, "success find one event");
+
+      if (!result) {
+        return response.notfound(res, "failed find one an event");
+      }
+
+      response.success(res, result, "success to find one an event");
     } catch (error) {
-      response.error(res, error, "failed find one event");
+      response.error(res, error, "failed to find one an event");
     }
   },
 
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notfound(res, "failed to update an event");
+      }
       const result = await EventModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
-      response.success(res, result, "success update an event");
+      response.success(res, result, "success to update an event");
     } catch (error) {
-      response.error(res, error, "failed update an event");
+      response.error(res, error, "failed to update an event");
     }
   },
 
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notfound(res, "failed to remove an event");
+      }
       const result = await EventModel.findByIdAndDelete(id, {
         new: true,
       });
-      response.success(res, result, "success remove an event");
+      response.success(res, result, "success to remove an event");
     } catch (error) {
-      response.error(res, error, "failed remove an event");
+      response.error(res, error, "failed to remove an event");
     }
   },
